Understanding the history so far
I've pushed previous commits in a hurry without proper documentation for personal decisions - the previous commits to this point should be ignored as "foundation building" testing.

As of now - (which cannot be said of previous commits) - using Node.js, the server successfully connects to MySQL to display the database entries, and the html page is acceptinging the CSS styling 🤕

This is the most functioning version and should be treated as "Start".